+++ metrics_utility/exceptions.py
@@ -1,0 +1,13 @@
+class BadShipTarget(Exception):
+    def __init__(self, message):
+        self.name = message
+
+
+class MissingRequiredEnvVar(Exception):
+    def __init__(self, message):
+        self.name = message
+
+
+class BadRequiredEnvVar(Exception):
+    def __init__(self, message):
+        self.name = message